internal/endpoints: Shutdown servers on shutdown signal
When closing a listener, the server won't accept any new connections.
However, we should still allow for the current connections / requests to
finish. Calling ``server.Shutdown`` allows us to wait for them to
finish. We'll be doing this only if drainConnectionsTimeout has been set
for the server.

Note that server.Serve always returns an error, and it will return
http.ErrServerClosed if server.Shutdown or server.Close was called, and
net.ErrClosed if the listener was closed.

Signed-off-by: Claudiu Belu <claudiu.belu@canonical.com>