internal/rest/resources: Shutdown servers after the request finishes
Note that in the case we fail to bootstrap / join the cluster, we'll be
resetting a few things, including the cluster membership. This includes
the HTTPS and unix socket servers we have open by closing them.

However, we cannot gracefully shutdown the servers, as there's at least
one connection that is still open: the bootstrap / join request. Forcing
the connection to close before we're able to write the request response
will result in the client getting an EOF error, and no information
regarding the failure.

Shutting down the server in a goroutine will address this issue: we'll be able
to return and write the HTTP response and then close the connection, finally
allowing the Servers to gracefully shutdown, and the clients to be happy.

Signed-off-by: Claudiu Belu <claudiu.belu@canonical.com>