internal/rest/resources: Start cleanup after request is done
When resetting a cluster member the shutdown procedures should be invoked only after the connection
that was used to trigger the reset has been closed (context cancelled).
The database might be already stopped but the request's context isn't yet cancelled.

Furthermore the state directory should be cleaned after all the listeners are closed (including inotify)
to prevent any errors trying to read the state directory.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>