internal/rest/resources: Continue forced member removal after db error
A forced member removal should remove the member in all places
(database, truststore etc.). This is done already on a best-effort base
to ensure that the cleanup happens on diverged state. For truststore and
member resets this is already the case.
This commit add this logic for the database removal.

Signed-off-by: Benjamin Schimke <benjamin.schimke@canonical.com>